+++ metric-anomaly-detector/src/main/java/org/hypertrace/alert/engine/metric/anomaly/detector/NotificationEventProducer.java
@@ -12,8 +12,12 @@ import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.hypertrace.alert.engine.metric.anomaly.datamodel.NotificationEvent;
 import org.hypertrace.alert.engine.metric.anomaly.datamodel.queue.KafkaConfigReader;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 class NotificationEventProducer {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(MetricAnomalyDetector.class);
 
   private final Producer<String, ByteBuffer> producer;
   private final KafkaConfigReader kafkaConfigReader;
@@ -26,10 +30,14 @@ class NotificationEventProducer {
     producer = new KafkaProducer<String, ByteBuffer>(properties);
   }
 
-  public void publish(NotificationEvent notificationEvent) throws IOException {
-    producer.send(
-        new ProducerRecord<String, ByteBuffer>(
-            properties.getProperty("topic"), null, notificationEvent.toByteBuffer()));
+  public void publish(NotificationEvent notificationEvent) {
+    try {
+      producer.send(
+          new ProducerRecord<String, ByteBuffer>(
+              properties.getProperty("topic"), null, notificationEvent.toByteBuffer()));
+    } catch (IOException e) {
+      LOGGER.error("Exception producing messages", e);
+    }
   }
 
   public void close() {
